Changed new content type notification in settings and updated tests
