+++ packages/js/src/settings/store/notifications.js
@@ -1,26 +1,9 @@
 import { createSlice, nanoid } from "@reduxjs/toolkit";
 import { get, omit } from "lodash";
-import apiFetch from "@wordpress/api-fetch";
 
 /**
  * @returns {Object} The initial state.
  */
-
-const NEW_CONTENT_ACTION_NAME = "removeNewContentNotification";
-
-/**
- * @returns {Object} Success or error action object.
- */
-export function* removeNewContentNotification() {
-	try {
-		yield{
-			type: NEW_CONTENT_ACTION_NAME,
-		};
-	} catch ( error ) {
-		// Empty.
-	}
-	return { type: "notifications/removeNotification", payload: "new-content-type" };
-}
 
 const slice = createSlice( {
 	name: "notifications",
@@ -57,16 +40,6 @@ export const notificationsSelectors = {
 	selectNotification: ( state, id ) => get( state, `notifications.${ id }`, null ),
 };
 
-export const removeNewContentNotificationControls = {
-	[ NEW_CONTENT_ACTION_NAME ]: async() => apiFetch( {
-		path: "/yoast/v1/needs-review/dismiss-notification",
-		method: "POST",
-	} ),
-};
-
-export const notificationsActions = {
-	...slice.actions,
-	removeNewContentNotification,
-};
+export const notificationsActions = slice.actions;
 
 export default slice.reducer;
